refactor: merging binding file
